address performance review followups
